requirements.txt
anyio==3.6.2
cachelib==0.10.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
cs50==9.2.5
decorator==5.1.1
fastapi==0.92.0
Flask==2.2.3
Flask-Session==0.4.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pydantic==1.10.5
python-dotenv==1.0.0
requests==2.28.2
sniffio==1.3.0
SQLAlchemy==1.4.46
sqlparse==0.4.3
starlette==0.25.0
termcolor==2.2.0
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0
validators==0.20.0
Werkzeug==2.2.3
